IotLabs: Fix issues with topology not being loaded well and some minor refactoring

+++ Experimentacion/IotLab/src/emulated_nodes/worker_node.h
@@ -21,7 +21,7 @@ public:
 	void handle_map_task();
 
 private:
-	int run_operation(int iterations);
+	int run_operation(long iterations);
 
 	std::string worker_ip;
 	std::string ip_to_coordinator;
